% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alProp.R
\docType{data}
\name{alProp}
\alias{alProp}
\title{Thermodynamic properties of Aluminum}
\format{
A dataframe with 5 variables and 599 observations.
\describe{
\item{T/K}{temperature}
\item{rho/(kg/m3)}{density}
\item{cp/(J/(kg.K))}{isobaric specific heat}
\item{K/(W/(m.K))}{thermal conductivity}
\item{alpha_thermal/(1/K)}{thermal expansion coefficient}
}
}
\usage{
alProp
}
\description{
This dataset contains temperature-dependent thermodynamic properties of Aluminum of 99.99\% purity.
It contains temperature, density, cp, thermal conductivity, and thermal expansion coefficient.
}
\keyword{datasets}
